refactor scene and scene manager
