+++ examples/src/scenes/CubeScene.h
@@ -10,7 +10,7 @@
 #ifndef CBIT_CUBESCENE_H
 #define CBIT_CUBESCENE_H
 
-#include "../../src/core/Scene.h"
+#include "core/project/Scene.h"
 #include "../../../src/core/ShaderProgram.h"
 #include "../../../src/core/camera/Camera.h"
