+++ examples/src/scenes/CubeTextureScene.h
@@ -11,7 +11,7 @@
 #ifndef CBIT_CUBETEXTURESCENE_H
 #define CBIT_CUBETEXTURESCENE_H
 
-#include "../../src/core/Scene.h"
+#include "core/project/Scene.h"
 #include "../../../src/core/ShaderProgram.h"
 #include "../../../src/core/camera/Camera.h"
 #include "../../../src/core/Texture.h"
