+++ examples/src/scenes/RectangleCameraScene.h
@@ -11,7 +11,7 @@
 #ifndef CBIT_RECTANGLECAMERASCENE_H
 #define CBIT_RECTANGLECAMERASCENE_H
 
-#include "../../src/core/Scene.h"
+#include "core/project/Scene.h"
 #include "../../../src/core/ShaderProgram.h"
 #include "../../../src/core/camera/Camera.h"
