+++ examples/src/scenes/SimpleScene.h
@@ -10,7 +10,7 @@
 #define SIMPLESCENE_H
 
 #include "../../../src/core/ecs/GameObject.h"
-#include "../../../src/core/Scene.h"
+#include "../../../src/core/project/Scene.h"
 
 class SimpleScene : public Scene {
 public:
