+++ src/core/project/Scene.h
@@ -11,9 +11,9 @@
 #ifndef CBIT_SCENE_H
 #define CBIT_SCENE_H
 
-#include "AssetManager.h"
-#include "input/Input.h"
-#include "ecs/EntityComponentSystem.h"
+#include "../AssetManager.h"
+#include "../input/Input.h"
+#include "../ecs/EntityComponentSystem.h"
 
 class Scene {
 public:
