+++ src/editor/Editor.h
@@ -14,7 +14,7 @@
 
 #include "EditorMainMenuBar.h"
 #include "ProfilePanel.h"
-#include "../core/SceneManager.h"
+#include "../core/project/SceneManager.h"
 #include "../core/camera/OrbitCamera.h"
 #include "../core/project/Project.h"
 #include "core/project/ProjectManager.h"
